coding/langchain/deepagent-langchain/Chapter_3_Customize/requirements.txt
deepagents
tavily-python
langchain-openai
langchain-anthropic
python-dotenv
pydantic
